Support DCA Staking on Pools
This adds support for staking DCA to become the "executor" which gives you the priviledge, but not obligation, to execute the distribution and earn the gelatoFeeShare. Gelato execution places a markup on the gas used (e.g., 50% on Optimism). The executor has the priviledge of executing the distribution for a markup less than Gelato (e.g., 49.99%).